Fix codespell arg
It turns out that -L lets you provide a list of words to ignore *on the
command line*. What we actually want is to provide a file with a list of
words to be ignored. The correct option for this is -I.

The reason we didn't notice is because the file is currently empty, but
it will be frustrating for whoever actually tries to add something to
it!